refactor(svelte): split codemod rewrite helpers from migrate-plot-props
Move AST walk, attribute→child emission, and import edits into
migrate-plot-props-rewrite.ts. Keep the public transform, result types,
and RULES table in migrate-plot-props.ts so package codemod exports stay
stable.
